requirement.txt
uvicorn
fastapi
python-dotenv
sqlalchemy
subprocess